Move core gem to pakyow namespace
