+++ pakyow-core/lib/pakyow/core/app_context.rb
