+++ pakyow-core/lib/pakyow/core/call_context.rb
@@ -1,9 +1,9 @@
-require 'core/helpers'
-require 'core/helpers/hooks'
-require 'core/router'
-require 'core/request'
-require 'core/response'
-require 'core/app_context'
+require 'pakyow/core/helpers'
+require 'pakyow/core/helpers/hooks'
+require 'pakyow/core/router'
+require 'pakyow/core/request'
+require 'pakyow/core/response'
+require 'pakyow/core/app_context'
 
 module Pakyow
   class CallContext
