+++ pakyow-core/lib/pakyow/core/config/app.rb
