+++ pakyow-core/lib/pakyow/core/config/cookies.rb
