+++ pakyow-core/lib/pakyow/core/config/session.rb
