+++ pakyow-core/lib/pakyow/core/errors.rb
