+++ pakyow-core/lib/pakyow/core/helpers.rb
