+++ pakyow-core/lib/pakyow/core/helpers/hooks.rb
