+++ pakyow-core/lib/pakyow/core/helpers/running.rb
