+++ pakyow-core/lib/pakyow/core/loader.rb
