+++ pakyow-core/lib/pakyow/core/middleware/logger.rb
