+++ pakyow-core/lib/pakyow/core/middleware/reloader.rb
