+++ pakyow-core/lib/pakyow/core/middleware/static.rb
@@ -1,5 +1,5 @@
-require 'core/config'
-require 'core/call_context'
+require 'pakyow/core/config'
+require 'pakyow/core/call_context'
 
 module Pakyow
   module Middleware
