+++ pakyow-core/lib/pakyow/core/multilog.rb
