+++ pakyow-core/lib/pakyow/core/route_template_defaults.rb
