+++ pakyow-core/spec/unit/config/session_config_spec.rb
@@ -1,5 +1,5 @@
 require_relative '../../spec_helper'
-require 'core/config/session'
+require 'pakyow/core/config/session'
 
 describe 'config.session' do
   it 'registers session config' do
